Better editing of double values
This is probably overcomplicated - using getInitialState instead of
initialState should allow us to make state just a String, then we can
remove some of the logic in render. However currently works well -
allows editing to invalid values, but only commits to cursor on valid
values, and allows new props to replace editing value whether it is
valid or not.