Extend Todo model, use router to determine which part of model is displayed
Todo model now has a TodoProject containing multiple lists, each
containing multiple todo items.
We parse the list and item index (if any) using router, and then use
this to construct a TodoPage that is passed to the view. It is wrapped
in a new Pages case class that also includes a callback to change the
page displayed. This allows for a single ServerRootComponent to
navigate through a TodoProject while retaining its state and so using
one websocket etc., supporting browser history.
There is quite a lot of boilerplate, but it should definitely be
possible to reduce this, particularly in DemoRoutes where we could
provide some helper methods, perhaps via implicits.