More flexible ListView
Can now change the type of P in CursorP when going from list-like type
to item type. This allows us to provide (for example) an "Actions"
object that can be used to do things on higher levels of the model, for
example a data type that contains a list of items can provide an action
to delete an item. Reimplemented old (and pointlessly strictly typed)
ListView apply with this approach.