Temporary fix for text/int/doubleView def problem
Will need to review everywhere for using defs that can lead us to
recreate components and then pass them props when rendering, instead of
passing new props to the same component instance.
Probably the only problematic ones are where we have changed a val to a
def to insert the `U` type parameter, however these are all obviously
unused, so we can replace with an existential type, or make U covariant
in deltas so we can have type Nothing where we don't actually create
Ids.